janitoring: Refactored "Motor" class.
 - refactored "motor.cpp".
 - renamed functions in "keywords.txt".
 - refactored all function names.
 - removed hardtabs.
